refactor: renamed recipe_scripts to recipe_parser
