+++ test/recipe_scripts_tests/kptncook_test.dart
@@ -3,7 +3,7 @@ import 'package:html/dom.dart';
 import 'package:ingredient_collector/src/models/ingredient.dart';
 import 'package:ingredient_collector/src/models/recipe_parsing_job.dart';
 import 'package:ingredient_collector/src/recipe_controller/recipe_controller.dart';
-import 'package:ingredient_collector/src/recipe_scripts/recipe_parser.dart'
+import 'package:ingredient_collector/src/recipe_parser/recipe_parser.dart'
     show KptnCookParser;
 
 import 'script_test_helper.dart';
